+++ gazebo/rendering/Scene.cc
@@ -18,6 +18,7 @@
 #include <functional>
 
 #include <boost/lexical_cast.hpp>
+#include <boost/make_shared.hpp>
 #include <ignition/math/Color.hh>
 #include <ignition/math/Helpers.hh>
 
@@ -162,6 +163,9 @@ Scene::Scene(const std::string &_name, const bool _enableVisualizations,
         &Scene::OnPoseMsg, this);
   }
   else
+  // When ready to use the direct API for updating scene poses from server,
+  // uncomment the following line and delete the if and else directly above
+  // if (!_isServer)
   {
     this->dataPtr->poseSub = this->dataPtr->node->Subscribe("~/pose/info",
         &Scene::OnPoseMsg, this);
@@ -204,6 +208,8 @@ Scene::Scene(const std::string &_name, const bool _enableVisualizations,
 //////////////////////////////////////////////////
 void Scene::Clear()
 {
+  this->dataPtr->initialized = false;
+
   this->dataPtr->connections.clear();
 
   this->dataPtr->poseSub.reset();
@@ -295,8 +301,6 @@ void Scene::Clear()
   this->dataPtr->skyxController = nullptr;
 
   RTShaderSystem::Instance()->RemoveScene(this->Name());
-
-  this->dataPtr->initialized = false;
 }
 
 //////////////////////////////////////////////////
@@ -2858,6 +2862,17 @@ void Scene::OnPoseMsg(ConstPosesStampedPtr &_msg)
     else
       this->dataPtr->poseMsgs.insert(std::make_pair(p.id(), p));
   }
+}
+
+/////////////////////////////////////////////////
+void Scene::UpdatePoses(const msgs::PosesStamped &_msg)
+{
+  auto msgptr = boost::make_shared<const msgs::PosesStamped>(_msg);
+  this->OnPoseMsg(msgptr);
+
+  std::unique_lock<std::mutex> lck(this->dataPtr->newPoseMutex);
+  this->dataPtr->newPoseAvailable = true;
+  this->dataPtr->newPoseCondition.notify_all();
 }
 
 /////////////////////////////////////////////////